Create save.php
While saving/publishing a file, your identity (username or id in the database ...) is encrypted using the alphapad permutation cipher. Then, your file F is saved and a list of trusted users for this specific file F is created. Next, both the ciphered id and hash of F are pushed in an array that is serialiazed in a string which is also hashed.The resulting hash is used to request a RFC3161 timestamp from a TSA via openssl and curl. The TSA's response is collected and the timestamp is extracted. At the end, we have: the data file F, the list of trusted users (i.e people who can access F without validation), the timestamp and the TSA response.